refactor: reduce remaining maintainability hotspots
PHASE B: Extract _CoordinatorConfigPropertiesMixin from ThesslaGreenModbusCoordinator
- Nine config-backed property pairs (host, port, slave_id, connection_type,
  connection_mode, serial_port, baud_rate, parity, stop_bits) moved from
  coordinator.py into new coordinator/config_properties.py
- coordinator.py non-empty: 666 → 605 (−61 lines)
- Coordinator package API (CoordinatorConfig, ThesslaGreenModbusCoordinator) unchanged

PHASE C: Promote scanner failure logging helpers to io_read_helpers
- mark_failed_addresses, log_read_abort, log_read_failure moved from
  scanner/io_read.py to scanner/io_read_helpers.py as public exports
- scanner/io_read.py non-empty: 714 → 701 (−13 lines)
- No HA imports introduced; scanner HA-independence preserved

PHASE E: Eliminate transport selection duplication in _write_holding_multi
- _write_holding_multi now delegates chunk writes to _write_registers_payload
  instead of repeating the three-branch transport selection pattern
- coordinator/schedule.py non-empty: 419 → 401 (−18 lines)
- Write/chunking/retry behavior unchanged; all 1938 tests pass

All maintained gates pass: ruff check/format, compileall, check_maintainability,
validate_entity_mappings (366 entities), pytest (1938 passed, 4 skipped).
